# Ms-Order/src/main/java/br/com/compass/pb/msorder/domain/dto/ItemDTO.java
package br.com.compass.pb.msorder.domain.dto;

import com.fasterxml.jackson.annotation.JsonFormat;
import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
import com.fasterxml.jackson.databind.annotation.JsonSerialize;
import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateDeserializer;
import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateSerializer;
import io.swagger.v3.oas.annotations.media.Schema;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Pattern;
import jakarta.validation.constraints.Positive;
import lombok.*;
import org.springframework.format.annotation.DateTimeFormat;

import java.math.BigDecimal;
import java.time.LocalDate;
@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class ItemDTO {
    @NotBlank
    @NotNull
    @Pattern(regexp = "^([a-zA-ZãÃéÉíÍóÓêÊôÔáÁ\s])+$", message =
            "O campo nome deve ser composto apenas por letras e não deve estar em branco!")
    @Schema(name = "name", example = "This is a name", type = "String")
    private String name;
    @Schema(name = "description", example = "this is a description", type = "String")
    private String description;
    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy")
    @JsonSerialize(using = LocalDateSerializer.class)
    @JsonDeserialize(using = LocalDateDeserializer.class)
    @Schema(name ="expirationDate", type = "String", example = "30-03-2023")
    private LocalDate expirationDate;
    @NotNull(message = "Campo obrigatório!")
    @Positive(message = "apenas valores positivos!")
    @Schema(name="value", example = "25", type = "Number")
    private BigDecimal value;

}