# Ms-Order/src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Microserviço Order
  description: Gerenciamento de pedidos
  contact:
    name: Jônatas Gomes
    url: https://github.com/Jonatas-Gomes
    email: jonatas.gomes2014@hotmail.com
  version: 0.1.0
...
servers:
  - url: http://localhost:8081/api
    description: URL Base da aplicação
paths:
...
  /pedidos/itens/{id}:
    patch:
      tags:
        - item-controller
      summary: Atualizar um item
      operationId: patchItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
components:
  schemas:
...
    ItemDTO:
      required:
        - name
        - value
      type: object
      properties:
        name:
          pattern: ^([a-zA-ZãÃéÉíÍóÓêÊôÔáÁ ])+$
          type: string
          example: This is a name
        description:
          type: string
          example: this is a description
        expirationDate:
          type: string
          format: date
          pattern: ^(3[01]|[12][0-9]|0[1-9])-(1[0-2]|0[1-9])-[0-9]{4}$
          example: '28-02-2023'
        value:
          type: number
          example: 25
...
    ItemResponse:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        value:
          type: number
        creationDate:
          type: string
          format: date
        expirationDate:
          type: string
          format: date
...